+++ spec/spicerack/root_object_spec.rb
@@ -26,7 +26,7 @@ RSpec.describe Spicerack::RootObject do
       context "with one event" do
         subject(:define_callbacks_with_handler) { example_class.__send__(:define_callbacks_with_handler, event) }
 
-        let(:event) { Faker::Internet.domain_word.to_sym }
+        let(:event) { Faker::Internet.domain_word.underscore.to_sym }
 
         it_behaves_like "callbacks and handler are defined for event"
       end
@@ -36,8 +36,8 @@ RSpec.describe Spicerack::RootObject do
           example_class.__send__(:define_callbacks_with_handler, event1, event2)
         end
 
-        let(:event1) { Faker::Internet.domain_word.to_sym }
-        let(:event2) { Faker::Internet.domain_word.to_sym }
+        let(:event1) { Faker::Internet.domain_word.underscore.to_sym }
+        let(:event2) { Faker::Internet.domain_word.underscore.to_sym }
 
         it_behaves_like "callbacks and handler are defined for event" do
           let(:expected_events) { [ event1, event2 ] }
@@ -61,7 +61,7 @@ RSpec.describe Spicerack::RootObject do
       context "with one event" do
         subject(:define_callbacks_with_handler) { example_class.__send__(:define_callbacks_with_handler, event) }
 
-        let(:event) { Faker::Internet.domain_word.to_sym }
+        let(:event) { Faker::Internet.domain_word.underscore.to_sym }
 
         it_behaves_like "callbacks and handler are defined for event"
       end
@@ -69,8 +69,8 @@ RSpec.describe Spicerack::RootObject do
       context "with several events" do
         subject(:define_callbacks_with_handler) { example_class.__send__(:define_callbacks_with_handler, event1, event2) }
 
-        let(:event1) { Faker::Internet.domain_word.to_sym }
-        let(:event2) { Faker::Internet.domain_word.to_sym }
+        let(:event1) { Faker::Internet.domain_word.underscore.to_sym }
+        let(:event2) { Faker::Internet.domain_word.underscore.to_sym }
 
         it_behaves_like "callbacks and handler are defined for event" do
           let(:expected_events) { [ event1, event2 ] }