+++ spicerack-styleguide/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v0.25.3
+
+- Add `rubocop-faker` to required gems ([#425](https://github.com/Freshly/spicerack/pull/425))
+
 ## v0.25.2
 
 *Release Date*: 10/16/2020